refactor conversation and chat debt surfaces
