+++ ui/src/components/ui/v0-ai-chat.tsx
@@ -1,38 +1,33 @@
 "use client";
 
-import { memo, useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { Streamdown } from "streamdown";
-import "streamdown/styles.css";
-import remarkGfm from "remark-gfm";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { AnimatePresence, motion } from "framer-motion";
 import {
   ArrowDown,
   ArrowUp,
   Brain,
   Check,
-  ChevronDown,
   Copy,
-  GraduationCap,
   Lightbulb,
   Loader2,
   Paperclip,
   Rocket,
   Sparkles,
-  Zap,
 } from "lucide-react";
 
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-// Shimmer now imported from ai-elements/reasoning
 import { Textarea } from "@/components/ui/textarea";
-import { cn } from "@/lib/utils";
-import { Reasoning, ReasoningTrigger, ReasoningContent } from "@/components/ai-elements/reasoning";
 import { ArtifactsProvider } from "@/components/workspace/artifacts/context";
+import { ChatLayout } from "@/components/workspace/chat-layout";
+import { MessageBubble, MessageListSkeleton, StreamingDots } from "@/components/workspace/chat-message-ui";
 import { ClarificationCard } from "@/components/workspace/messages/clarification-card";
-import { FilesIcon, XIcon } from "lucide-react";
+import { ModePicker, ModelPicker } from "@/components/workspace/chat-controls";
 import { SubtasksProvider, useUpdateSubtask, useSubtaskContext } from "@/core/tasks/context";
 import { SubtaskCard } from "@/components/workspace/messages/subtask-card";
 import { parseClarificationContent } from "@/core/messages/clarification";
+import { cn } from "@/lib/utils";
+import { XIcon } from "lucide-react";
 import { toast } from "sonner";
 
 interface V0ChatProps {
@@ -147,55 +142,8 @@ const QUICK_PROMPTS = [
 const MODEL_FETCH_RETRY_COUNT = 3;
 const MODEL_FETCH_RETRY_DELAY_MS = 400;
 
-const MODE_OPTIONS: {
-  id: ConversationMode;
-  label: string;
-  description: string;
-  accentClassName: string;
-  icon: typeof Zap;
-}[] = [
-  {
-    id: "flash",
-    label: "Flash",
-    description: "最快回复，不展开推理",
-    accentClassName: "border-[var(--status-running-border)] bg-[var(--status-running-bg)] text-[var(--status-running)]",
-    icon: Zap,
-  },
-  {
-    id: "thinking",
-    label: "Thinking",
-    description: "保留推理过程，单轮深入分析",
-    accentClassName: "border-[var(--status-chat-border)] bg-[var(--status-chat-bg)] text-[var(--status-chat)]",
-    icon: Lightbulb,
-  },
-  {
-    id: "pro",
-    label: "Pro",
-    description: "先规划再执行",
-    accentClassName: "border-[var(--status-done-border)] bg-[var(--status-done-bg)] text-[var(--status-done)]",
-    icon: GraduationCap,
-  },
-  {
-    id: "ultra",
-    label: "Ultra",
-    description: "启用完整协作流程",
-    accentClassName: "border-[var(--status-approval-border)] bg-[var(--status-approval-bg)] text-[var(--status-approval)]",
-    icon: Rocket,
-  },
-];
-
 const DEFAULT_MODE: ConversationMode = "flash";
 const DEFAULT_MODEL = "";
-
-const streamAnimateOptions = {
-  animation: "fadeIn" as const,
-  duration: 300,
-  stagger: 20,
-  sep: "word" as const,
-};
-
-const streamingRemarkPlugins = [remarkGfm];
-const staticRemarkPlugins = [remarkGfm];
 
 function generateId() {
   return Math.random().toString(36).slice(2, 10);
@@ -276,389 +224,6 @@ function statusLabel(phase: RuntimeState["phase"]) {
   if (phase === "completed") return "已完成";
   if (phase === "error") return "失败";
   return "待开始";
-}
-
-function ModePicker({
-  selected,
-  onSelect,
-}: {
-  selected: ConversationMode;
-  onSelect: (id: ConversationMode) => void;
-}) {
-  const [open, setOpen] = useState(false);
-  const current =
-    MODE_OPTIONS.find((mode) => mode.id === selected) ?? MODE_OPTIONS[0];
-  const CurrentIcon = current.icon;
-
-  return (
-    <div className="relative">
-      <button
-        type="button"
-        onClick={() => { setOpen((prev) => !prev); }}
-        aria-label={`当前执行模式：${current.label}`}
-        className={cn(
-          "group flex min-h-10 items-center gap-2 rounded-lg border px-3 py-2 text-left transition-all duration-200 hover:border-[var(--warm-border)] focus-visible:border-[var(--warm-ring)] focus-visible:ring-2 focus-visible:ring-[var(--warm-ring)]/50",
-          current.accentClassName,
-        )}
-      >
-        <span className="flex size-6 items-center justify-center rounded-md border border-black/5 bg-[var(--warm-ivory)]">
-          <CurrentIcon className="size-3" />
-        </span>
-        <span className="min-w-0">
-          <span className="block text-[10px] leading-4 font-semibold tracking-[0.08em]">
-            {current.label}
-          </span>
-        </span>
-        <ChevronDown
-          className={cn(
-            "size-3 shrink-0 transition-transform duration-200",
-            open && "rotate-180",
-          )}
-        />
-      </button>
-
-      <AnimatePresence>
-        {open && (
-          <>
-            <div
-              className="fixed inset-0 z-40"
-              onClick={() => { setOpen(false); }}
-            />
-            <motion.div
-              initial={{ opacity: 0, y: 8, scale: 0.97 }}
-              animate={{ opacity: 1, y: 0, scale: 1 }}
-              exit={{ opacity: 0, y: 8, scale: 0.97 }}
-              transition={{
-                type: "spring",
-                stiffness: 360,
-                damping: 28,
-                mass: 0.9,
-              }}
-              className="absolute bottom-full left-0 z-50 mb-2.5 w-[286px] rounded-[18px] border border-border bg-card p-2"
-            >
-              <div className="mb-1.5 px-1">
-                <p className="text-[10px] uppercase tracking-[0.1em] text-muted-foreground">
-                  执行模式
-                </p>
-                <p className="text-[12px] text-foreground">
-                  选择这轮临时会话的执行方式
-                </p>
-              </div>
-              <div className="space-y-1.5">
-                {MODE_OPTIONS.map((mode, index) => {
-                  const Icon = mode.icon;
-                  const isSelected = mode.id === selected;
-
-                  return (
-                    <motion.button
-                      key={mode.id}
-                      type="button"
-                      onClick={() => {
-                        onSelect(mode.id);
-                        setOpen(false);
-                      }}
-                      initial={{ opacity: 0, y: 6 }}
-                      animate={{ opacity: 1, y: 0 }}
-                      transition={{ delay: index * 0.03 }}
-                      className={cn(
-                        "flex w-full items-start gap-2.5 rounded-[14px] border px-3 py-2.5 text-left transition-colors",
-                        isSelected
-                          ? cn("bg-card", mode.accentClassName)
-                          : "border-[var(--warm-border)] bg-[var(--warm-ivory)] text-[var(--neutral-900)] hover:border-[var(--warm-ring)] hover:bg-[var(--neutral-150)]",
-                      )}
-                    >
-                      <span
-                        className={cn(
-                          "mt-0.5 flex size-8 shrink-0 items-center justify-center rounded-md border bg-background",
-                          isSelected ? "border-black/5" : "border-border/80",
-                        )}
-                      >
-                        <Icon className="size-3.5" />
-                      </span>
-                      <span className="min-w-0 flex-1">
-                        <span className="text-[12px] font-semibold">
-                          {mode.label}
-                        </span>
-                        <span className="mt-0.5 block text-[11px] leading-4 opacity-80">
-                          {mode.description}
-                        </span>
-                      </span>
-                    </motion.button>
-                  );
-                })}
-              </div>
-            </motion.div>
-          </>
-        )}
-      </AnimatePresence>
-    </div>
-  );
-}
-
-function ModelPicker({
-  models,
-  selected,
-  onSelect,
-  isLoading,
-  loadError,
-  onRetry,
-}: {
-  models: RuntimeModelOption[];
-  selected: string;
-  onSelect: (id: string) => void;
-  isLoading: boolean;
-  loadError: string | null;
-  onRetry: () => void;
-}) {
-  const [open, setOpen] = useState(false);
-  const current = models.find((model) => model.name === selected) ?? models[0];
-  const currentLabel =
-    current?.display_name ||
-    current?.name ||
-    (isLoading ? "加载模型..." : loadError ? "模型加载失败" : "未配置模型");
-  const isDisabled = isLoading || (!loadError && models.length <= 1);
-
-  return (
-    <div className="relative">
-      <button
-        type="button"
-        onClick={() => {
-          if (loadError) {
-            onRetry();
-            return;
-          }
-          if (!isDisabled) {
-            setOpen((prev) => !prev);
-          }
-        }}
-        disabled={isDisabled}
-        title={loadError ?? undefined}
-        className="flex min-h-10 items-center gap-2 rounded-lg border border-transparent bg-transparent px-3 text-[11px] tracking-[0.06em] text-[var(--neutral-600)] transition-all duration-200 hover:border-[var(--warm-border)] hover:bg-[var(--neutral-150)] hover:text-[var(--neutral-800)] focus-visible:border-[var(--warm-ring)] focus-visible:ring-2 focus-visible:ring-[var(--warm-ring)]/50"
-      >
-        <Sparkles className="size-3.5" />
-        <span className="max-w-[140px] truncate">{currentLabel}</span>
-      </button>
-
-      <AnimatePresence>
-        {open && (
-          <>
-            <div
-              className="fixed inset-0 z-40"
-              onClick={() => { setOpen(false); }}
-            />
-            <motion.div
-              initial={{ opacity: 0, y: 4, scale: 0.96 }}
-              animate={{ opacity: 1, y: 0, scale: 1 }}
-              exit={{ opacity: 0, y: 4, scale: 0.96 }}
-              transition={{
-                type: "spring",
-                stiffness: 500,
-                damping: 30,
-                mass: 0.8,
-              }}
-              className="absolute bottom-full left-0 z-50 mb-2 w-[220px] overflow-hidden rounded-[16px] border border-border bg-card p-1.5"
-            >
-              {models.map((model) => (
-                <button
-                  key={model.name}
-                  type="button"
-                  onClick={() => {
-                    onSelect(model.name);
-                    setOpen(false);
-                  }}
-                  className={cn(
-                    "flex min-h-11 w-full items-center gap-2 rounded-[14px] px-3 py-2 text-[13px] transition-colors",
-                    model.name === selected
-                      ? "bg-secondary text-foreground font-medium"
-                      : "text-muted-foreground hover:bg-secondary hover:text-foreground",
-                  )}
-                >
-                  <Sparkles className="size-3.5 shrink-0" />
-                  <span className="truncate">
-                    {model.display_name || model.name}
-                  </span>
-                </button>
-              ))}
-            </motion.div>
-          </>
-        )}
-      </AnimatePresence>
-    </div>
-  );
-}
-
-const MessageBubble = memo(function MessageBubble({
-  message,
-  isMessageStreaming = false,
-}: {
-  message: ChatMessage;
-  isMessageStreaming?: boolean;
-}) {
-  const isUser = message.role === "user";
-  const hasCodeBlock = !isUser && message.content.includes("```");
-  const shouldShowMessageCopy =
-    message.content.trim().length > 0 && (isUser || !hasCodeBlock);
-  const [copied, setCopied] = useState(false);
-
-  useEffect(() => {
-    if (!copied) {
-      return;
-    }
-
-    const timer = window.setTimeout(() => { setCopied(false); }, 2000);
-    return () => { window.clearTimeout(timer); };
-  }, [copied]);
-
-  const handleCopy = useCallback(async () => {
-    if (!message.content.trim()) {
-      return;
-    }
-
-    try {
-      await navigator.clipboard.writeText(message.content);
-      setCopied(true);
-    } catch (error) {
-      console.error("Failed to copy message content:", error);
-    }
-  }, [message.content]);
-
-  return (
-    <div
-      className={cn(
-        "group flex w-full",
-        isUser ? "justify-end" : "justify-start md:pr-8",
-      )}
-    >
-      <div
-        className={cn(
-          "relative max-w-[90%]",
-          isUser
-            ? "rounded-[22px] border user-bubble px-[20px] py-[14px] text-[var(--neutral-900)]"
-            : "px-1 py-1 md:px-2",
-        )}
-      >
-        {shouldShowMessageCopy ? (
-          <div
-            className={cn(
-              "absolute transition-opacity md:opacity-0 md:group-hover:opacity-100 md:group-focus-within:opacity-100",
-              isUser
-                ? "-right-1 -top-1 opacity-100"
-                : "right-2 top-1 opacity-80",
-            )}
-          >
-            <Button
-              type="button"
-              variant="ghost"
-              size="icon-xs"
-              onClick={() => {
-                void handleCopy();
-              }}
-              className={cn(
-                "h-8 w-8 rounded-[10px] text-muted-foreground md:h-7 md:w-7",
-                isUser
-                  ? "border border-border bg-card hover:bg-card hover:text-foreground"
-                  : "border border-transparent bg-[var(--warm-ivory)] hover:border-[var(--warm-border)] hover:bg-[var(--neutral-150)] hover:text-[var(--neutral-700)]",
-              )}
-              title={copied ? "已复制" : "复制消息"}
-            >
-              {copied ? (
-                <Check className="size-3 text-[var(--status-done)]" />
-              ) : (
-                <Copy className="size-3" />
-              )}
-            </Button>
-          </div>
-        ) : null}
-
-        {!isUser && message.thinking ? (
-          <Reasoning
-            isStreaming={message.isReasoningStreaming}
-            defaultOpen={true}
-            className="mb-4"
-          >
-            <ReasoningTrigger />
-            <ReasoningContent>{message.thinking}</ReasoningContent>
-          </Reasoning>
-        ) : null}
-
-        {message.content ? (
-          isUser ? (
-            <div className="whitespace-pre-wrap font-[var(--font-body)] text-[14px] leading-[22px]">
-              {message.content}
-            </div>
-          ) : (
-            <div className="assistant-markdown prose prose-sm max-w-none px-[12px] py-1 font-[var(--font-body)] text-[14px] leading-[24px] tracking-[-0.003em] text-foreground prose-headings:font-sans prose-pre:my-0 prose-pre:mx-0 prose-pre:rounded-none prose-pre:border-0 prose-pre:bg-transparent prose-pre:p-0 prose-code:font-mono">
-              <Streamdown
-                mode={isMessageStreaming ? "streaming" : "static"}
-                remarkPlugins={
-                  isMessageStreaming
-                    ? streamingRemarkPlugins
-                    : staticRemarkPlugins
-                }
-                animated={isMessageStreaming ? streamAnimateOptions : false}
-              >
-                {message.content}
-              </Streamdown>
-            </div>
-          )
-        ) : (
-          <div className="flex items-center gap-2 text-[14px] leading-[22px] text-muted-foreground">
-            <StreamingDots />
-            正在整理回复
-          </div>
-        )}
-      </div>
-    </div>
-  );
-});
-
-function StreamingDots({ className }: { className?: string }) {
-  return (
-    <div
-      className={cn("inline-flex items-center", className)}
-      aria-hidden="true"
-    >
-      {[0, 1, 2].map((index) => (
-        <span
-          key={index}
-          className="chat-stream-dot"
-          style={{ animationDelay: `${index * 0.16}s` }}
-        />
-      ))}
-    </div>
-  );
-}
-
-function MessageListSkeleton() {
-  return (
-    <div className="mx-auto flex w-full max-w-[760px] flex-col gap-7 px-6 py-6">
-      <div className="flex justify-end">
-        <div className="w-full max-w-[420px] space-y-2">
-          <div className="skeleton-line h-4 rounded-full" />
-          <div className="skeleton-line ml-auto h-4 w-[72%] rounded-full" />
-        </div>
-      </div>
-
-      <div className="flex justify-start">
-        <div className="w-full max-w-[560px] space-y-2">
-          <div className="skeleton-line h-4 rounded-full" />
-          <div className="skeleton-line h-4 w-[94%] rounded-full" />
-          <div className="skeleton-line h-4 w-[68%] rounded-full" />
-        </div>
-      </div>
-
-      <div className="flex justify-start">
-        <div className="w-full max-w-[520px] rounded-2xl border border-[var(--warm-border)] bg-[var(--neutral-150)] px-4 py-4">
-          <div className="space-y-2">
-            <div className="skeleton-line h-3.5 w-[120px] rounded-full" />
-            <div className="skeleton-line h-4 rounded-full" />
-            <div className="skeleton-line h-4 w-[88%] rounded-full" />
-            <div className="skeleton-line h-4 w-[54%] rounded-full" />
-          </div>
-        </div>
-      </div>
-    </div>
-  );
 }
 
 // Internal component that uses useUpdateSubtask (must be inside SubtasksProvider)
@@ -1941,158 +1506,5 @@ export function V0Chat(props: V0ChatProps) {
     <SubtasksProvider>
       <V0ChatInner {...props} />
     </SubtasksProvider>
-  );
-}
-
-// ============================================================================
-// ChatLayout - Two-panel layout with Artifacts (Simplified from DeerFlow)
-// ============================================================================
-
-interface ChatLayoutProps {
-  children: React.ReactNode;
-  conversationId?: string;
-  artifacts: string[];
-  selectedArtifact: string | null;
-  onSelectArtifact: (path: string | null) => void;
-  artifactsOpen: boolean;
-  setArtifactsOpen: (open: boolean) => void;
-}
-
-function ChatLayout({
-  children,
-  conversationId,
-  artifacts,
-  selectedArtifact,
-  onSelectArtifact,
-  artifactsOpen,
-  setArtifactsOpen,
-}: ChatLayoutProps) {
-  return (
-    <div className="flex h-full w-full">
-      {/* Left Panel - Chat */}
-      <div className={cn(
-        "relative flex flex-col h-full transition-all duration-300",
-        artifactsOpen ? "w-[60%]" : "w-full"
-      )}>
-        {artifacts.length > 0 && !artifactsOpen && (
-          <div className="absolute top-3 right-3 z-10">
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => setArtifactsOpen(true)}
-              className="gap-1.5 text-xs"
-            >
-              <FilesIcon className="size-3.5" />
-              {artifacts.length} 个产物
-            </Button>
-          </div>
-        )}
-        {children}
-      </div>
-
-      {/* Resize Handle - Only show when artifacts panel is open */}
-      {artifactsOpen && (
-        <div className="w-px bg-border hover:bg-border/80 cursor-col-resize" />
-      )}
-
-      {/* Right Panel - Artifacts */}
-      {artifactsOpen && (
-        <div className="w-[40%] min-w-[300px] max-w-[600px] bg-background border-l flex flex-col h-full">
-          {/* Artifacts Header */}
-          <div className="flex items-center justify-between border-b px-4 py-3">
-            <h3 className="font-medium text-sm">Artifacts</h3>
-            <div className="flex items-center gap-2">
-              {artifacts.length > 0 && (
-                <span className="text-muted-foreground text-xs">
-                  {artifacts.length} 个文件
-                </span>
-              )}
-              <Button
-                variant="ghost"
-                size="icon-sm"
-                onClick={() => { setArtifactsOpen(false); }}
-              >
-                <XIcon className="size-4" />
-              </Button>
-            </div>
-          </div>
-
-          {/* Artifacts Content */}
-          <div className="flex-1 overflow-hidden">
-            {artifacts.length === 0 ? (
-              <div className="flex h-full flex-col items-center justify-center p-8 text-center">
-                <FilesIcon className="size-12 text-muted-foreground/50 mb-4" />
-                <h4 className="text-sm font-medium mb-2">没有 Artifacts</h4>
-                <p className="text-muted-foreground text-xs max-w-[200px]">
-                  当 AI 生成文件时，它们将显示在这里
-                </p>
-              </div>
-            ) : selectedArtifact ? (
-              <div className="h-full flex flex-col">
-                <div className="flex items-center justify-between border-b px-4 py-2 bg-muted/50">
-                  <span className="text-sm font-medium truncate max-w-[200px]">
-                    {selectedArtifact.split('/').pop()}
-                  </span>
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    onClick={() => { onSelectArtifact(null); }}
-                  >
-                    返回列表
-                  </Button>
-                </div>
-                <div className="flex-1 overflow-auto p-4">
-                  <iframe
-                    src={`/api/conversations/${conversationId}/artifacts${selectedArtifact}`}
-                    className="w-full h-full border rounded-lg"
-                    title={selectedArtifact}
-                  />
-                </div>
-              </div>
-            ) : (
-              <div className="p-4">
-                <ArtifactFileListSimple
-                  files={artifacts}
-                  onSelect={onSelectArtifact}
-                />
-              </div>
-            )}
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// ============================================================================
-// Simplified Artifact File List for ChatLayout
-// ============================================================================
-
-interface ArtifactFileListSimpleProps {
-  files: string[];
-  onSelect: (path: string) => void;
-}
-
-function ArtifactFileListSimple({ files, onSelect }: ArtifactFileListSimpleProps) {
-  return (
-    <div className="flex flex-col gap-2">
-      {files.map((file) => (
-        <button
-          key={file}
-          onClick={() => { onSelect(file); }}
-          className="flex items-center gap-3 rounded-lg border p-3 text-left hover:bg-muted transition-colors"
-        >
-          <FilesIcon className="size-5 text-muted-foreground" />
-          <div className="flex-1 min-w-0">
-            <p className="text-sm font-medium truncate">
-              {file.split('/').pop()}
-            </p>
-            <p className="text-xs text-muted-foreground truncate">
-              {file}
-            </p>
-          </div>
-        </button>
-      ))}
-    </div>
   );
 }
